pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bunq-hackathon"
version = "0.1.0"
description = "Multi-modal AI financial assistant for bunq Hackathon 7.0"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "openai>=1.50.0",
    "bunq-sdk>=1.14.9",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "Pillow>=10.4.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "websockets>=13.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]
offline = [
    "faster-whisper>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"
